requirements.txt
appdirs==1.4.3
chardet==3.0.2
Django==1.11
django-filter==1.0.2
djangorestframework==3.6.2
Jinja2==2.9.6
Markdown==2.6.8
MarkupSafe==1.0
packaging==16.8
pyparsing==2.2.0
PyPDF2==1.26.0
pytz==2017.2
six==1.10.0
